requirements.txt
openai
streamlit
python-dotenv

streamlit-chat

langchain

